put review upload in util file, added fields to review page
