+++ bruinrent/src/components/screens/ReviewPage.js
@@ -19,7 +19,6 @@ const ReviewPage = ({ addReview }) => {
         overall: 0,
         cleanliness: 0,
         noise: 0,
-        social: 0,
         landlord: 0,
         location: 0,
     });
@@ -29,6 +28,13 @@ const ReviewPage = ({ addReview }) => {
     const { user } = useAuthContext();
     const [review, setReview] = useState("");
     const [address, setAddress] = useState("");
+    const [apartmentName, setApartmentName] = useState("");
+    const [totalMonthlyRent, setTotalMonthlyRent] = useState(null);
+    const [residentName, setResidentName] = useState("");
+    const [residentEmail, setResidentEmail] = useState("");
+    const [isAnonymous, setIsAnonymous] = useState(false);
+    const [beds, setBeds] = useState(null);
+    const [baths, setBaths] = useState(null);
 
     // const [firstName, setFirstName] = useState("");
     // const [lastName, setLastName] = useState("");
@@ -53,31 +59,6 @@ const ReviewPage = ({ addReview }) => {
             [category]: newRating,
         }));
     };
-
-    // const handleSubmit = async (e) => {
-    //     e.preventDefault();
-    //     const longLat = await addressToLongLat(address);
-    //     const latLong = [longLat[1], longLat[0]];
-    //     const formData = {
-    //         address: address,
-    //         latLong: latLong,
-    //         review: review,
-    //         rating: rating,
-    //     };
-
-    //     const collectionRef = collection(firestore, `users/${user.uid}/reviews`);
-
-    //     try {
-    //         const docRef = await addDoc(collectionRef, formData);
-    //         console.log("Document written with ID: ", docRef.id);
-    //     } catch (error) {
-    //         console.error("Error adding document: ", error);
-    //     }
-
-    //     navigate("/");
-
-    //     console.log(formData);
-    // };
 
     const handleSubmit = async (e) => {
         e.preventDefault();
@@ -110,7 +91,7 @@ const ReviewPage = ({ addReview }) => {
         } catch (error) {
             console.error("Error:", error.message);
 
-            alert("Error: Invalid address. Please enter a valid address.");
+            // alert("Error: Invalid address. Please enter a valid address.");
         }
     };
 
@@ -142,10 +123,42 @@ const ReviewPage = ({ addReview }) => {
                                 type="address"
                                 value={address}
                                 onChange={(e) => setAddress(e.target.value)}
-                                // value={addressDesc}
-                                // onChange={(e) => setAddressDesc(e.target.value)}
-                                //onChange={onSearch}
                             />
+                            <text className="title-text">Apartment Name</text>
+                            <textarea
+                                className="address-review-text"
+                                type="apartmentName"
+                                value={apartmentName}
+                                onChange={(e) => setApartmentName(e.target.value)}
+                            />
+                            <text className="title-text">Apartment Info</text>
+                                <label> Total Monthly Rent: 
+                                    <input
+                                        className="address-review-text"
+                                        type="text"
+                                        pattern="[0-9]*"
+                                        value={totalMonthlyRent}
+                                        onChange={(e) => setTotalMonthlyRent(e.target.value.replace(/[^0-9]/ig, ""))}
+                                    />
+                                </label>
+                                <label> Beds
+                                    <input
+                                        className="address-review-text"
+                                        type="text"
+                                        pattern="[0-9]*"
+                                        value={beds}
+                                        onChange={(e) => setBeds(e.target.value.replace(/[^0-9]/ig, ""))}
+                                    />
+                                </label>
+                                <label> Baths 
+                                    <input
+                                        className="address-review-text"
+                                        type="text"
+                                        pattern="[0-9]*"
+                                        value={baths}
+                                        onChange={(e) => setBaths(e.target.value.replace(/[^0-9]/ig, ""))}
+                                    />
+                                </label>
                             <text className="title-text">Resident Info</text>
                             <div className="resident-info-year">
                                 <label for="resident-year">Your Year:</label>
@@ -167,6 +180,28 @@ const ReviewPage = ({ addReview }) => {
                                         Super Senior
                                     </option>
                                 </select>
+                                <label for="resident-name">Your name</label>
+                                <input
+                                    className="address-review-text"
+                                    type="text"
+                                    value={residentName}
+                                    onChange={(e) => setResidentName(e.target.value)}
+                                />
+                                <label for="resident-email">Email</label>
+                                <input
+                                    className="address-review-text"
+                                    type="text"
+                                    value={residentEmail}
+                                    onChange={(e) => setResidentEmail(e.target.value)}
+                                />
+                                <label>
+                                    Anonymous?
+                                    <input
+                                    type="checkbox"
+                                    checked={isAnonymous}
+                                    onChange={() => setIsAnonymous(!isAnonymous)}
+                                    />
+                                </label>
                             </div>
                             <text className="title-text">Write a Review</text>
                             <textarea
@@ -198,13 +233,6 @@ const ReviewPage = ({ addReview }) => {
                             title="Noise"
                             onRatingChange={(newRating) =>
                                 handleRatingChange("noise", newRating)
-                            }
-                        />
-                        <RatingStars
-                            id="social"
-                            title="Social"
-                            onRatingChange={(newRating) =>
-                                handleRatingChange("social", newRating)
                             }
                         />
                         <RatingStars
